Regroup form inputs by row
Grouping the inputs by row makes for a more logical
sequence of inputs on smaller screen sizes. Smaller
screen sizes show the inputs stacked on top of each
other and so when using the column sequence, the first
name input appears above the email input (and so on),
which is awkward when the user gets to filling in
the last name input. Regrouping the inputs by row
solves this issue.